Add list_commands.sh script to list all available aliases and functions
This commit adds a new script called list_commands.sh, which lists all available aliases and functions provided by the alias library. The script checks if the aliases are loaded by verifying the existence of the 'list_aliases' function. If the aliases are not loaded, it displays an error message and exits. Otherwise, it calls the list_aliases function to list all the available aliases and functions.